requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
SQLAlchemy==2.0.30
aiosqlite==0.20.0
